requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Environment variables
python-dotenv==1.0.1

# HTTP requests
requests==2.31.0

# OpenAI (for LLM summaries)
openai==1.10.0

# Anthropic (alternative LLM)
anthropic==0.15.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0